move: guideStore 파일 이름 변경, 폴더 이동
